Fix warnings

+++ include/sha1.hpp
@@ -81,13 +81,16 @@ namespace sha1
         uint32_t digest[DIGEST_INTS] = {0x67452301, 0xEFCDAB89, 0x98BADCFE, 0x10325476, 0xC3D2E1F0};
         std::vector<uint8_t> buffer;
         uint32_t block[BLOCK_INTS];
-        int b = 0;
+        uint32_t b = 0;
         for (b = 0; b + BLOCK_BYTES <= s.size(); b += BLOCK_BYTES)
         {
             buffer.assign(s.begin() + b, s.begin() + b + BLOCK_BYTES);
 
             for (uint32_t i = 0; i < BLOCK_INTS; i++)
-                block[i] = buffer[4 * i + 3] | buffer[4 * i + 2] << 8 | buffer[4 * i + 1] << 16 | buffer[4 * i + 0] << 24;
+                block[i] = static_cast<uint32_t>(buffer[4 * i + 3] |
+                                                 buffer[4 * i + 2] << 8 |
+                                                 buffer[4 * i + 1] << 16 |
+                                                 buffer[4 * i + 0] << 24);
 
             transform(block, digest);
         }
@@ -99,7 +102,10 @@ namespace sha1
             buffer.push_back(0x00);
 
         for (uint32_t i = 0; i < BLOCK_INTS; i++)
-            block[i] = buffer[4 * i + 3] | buffer[4 * i + 2] << 8 | buffer[4 * i + 1] << 16 | buffer[4 * i + 0] << 24;
+            block[i] = static_cast<uint32_t>(buffer[4 * i + 3] |
+                                             buffer[4 * i + 2] << 8 |
+                                             buffer[4 * i + 1] << 16 |
+                                             buffer[4 * i + 0] << 24);
 
         if (origSize > BLOCK_BYTES - 8)
         {
